requirements.txt
Flask==2.2.2
Flask_Bootstrap==3.3.7.1
numpy==1.21.6
Pillow==9.4.0
SQLAlchemy==1.4.46
